+++ crates/macros/src/config/field.rs
@@ -131,11 +131,9 @@ impl<'l> Field<'l> {
             };
 
             stmts.push(quote! {
-                if let Err(error) = #func(setting, self, context, finalize) {
-                    errors.push(schematic::ValidateErrorType::setting(
-                        path.join_key(#key_quoted),
-                        error,
-                    ));
+                if let Err(mut error) = #func(setting, self, context, finalize) {
+                    error.prepend_path(path.join_key(#key_quoted));
+                    errors.push(error);
                 }
             });
         }
@@ -157,9 +155,9 @@ impl<'l> Field<'l> {
         let second = if self.is_required() {
             quote! {
                 if finalize && self.#key.is_none() {
-                    errors.push(schematic::ValidateErrorType::setting_required(
-                        path.join_key(#key_quoted),
-                    ));
+                    let mut error = schematic::ValidateError::required();
+                    error.prepend_path(path.join_key(#key_quoted));
+                    errors.push(error);
                 }
             }
         } else {